Fixing tests (patching MongoClient, moving test classes out of __init__ files)
